move data_esets mapping suggestings to import_files mapping suggestions
